moved some comm functionality to its module
