+++ bot/modules/community/commands.js
@@ -3,7 +3,7 @@
 const logger = require('../../../logger');
 const messages = require('../../messages');
 const { Community, Order } = require('../../../models');
-const { validateUser, validateParams } = require('../../validations');
+const { validateParams } = require('../../validations');
 
 async function getOrderCountByCommunity() {
   const data = await Order.aggregate([
@@ -24,11 +24,56 @@ async function findCommunities(currency) {
   });
 }
 
+exports.setComm = async ctx => {
+  try {
+    const { user } = ctx;
+
+    const [groupName] = await validateParams(
+      ctx,
+      2,
+      '\\<_@communityGroupName / off_\\>'
+    );
+    if (!groupName) {
+      return;
+    }
+
+    if (groupName === 'off') {
+      user.default_community_id = null;
+      await user.save();
+      await messages.noDefaultCommunityMessage(ctx);
+      return;
+    }
+    // Allow find communities case insensitive
+    const regex = new RegExp(['^', groupName, '$'].join(''), 'i');
+    const community = await Community.findOne({ group: regex });
+    if (!community) {
+      await messages.communityNotFoundMessage(ctx);
+      return;
+    }
+
+    user.default_community_id = community._id;
+    await user.save();
+
+    await messages.operationSuccessfulMessage(ctx);
+  } catch (error) {
+    logger.error(error);
+  }
+};
+
+exports.myComms = async ctx => {
+  try {
+    const user = ctx.user;
+
+    const communities = await Community.find({ creator_id: user._id });
+
+    await messages.showUserCommunitiesMessage(ctx, communities);
+  } catch (error) {
+    logger.error(error);
+  }
+};
+
 exports.findCommunity = async ctx => {
   try {
-    const user = await validateUser(ctx, false);
-    if (!user) return;
-
     const [fiatCode] = await validateParams(ctx, 2, '\\<_fiat code_\\>');
     if (!fiatCode) return;
